use shoebox messaging endpoint for ext reply actions
