+++ kifi-backend/modules/eliza/app/com/keepit/eliza/commanders/EmailMessageProcessingCommander.scala
@@ -1,17 +1,15 @@
 package com.keepit.eliza.commanders
 
 import com.google.inject.Inject
+import com.keepit.discussion.MessageSource
 import com.keepit.eliza.model._
 import com.keepit.common.logging.Logging
 import com.kifi.franz.SQSQueue
 import com.keepit.eliza.mail.MailNotificationReply
-import com.keepit.common.db.Id
-import scala.Some
 import com.keepit.eliza.model.NonUserThread
 import com.keepit.common.healthcheck.AirbrakeNotifier
 import play.api.libs.concurrent.Execution.Implicits.defaultContext
 import scala.util.{ Success, Failure }
-import com.keepit.common.db.slick.Database
 import com.keepit.heimdal.HeimdalContextBuilderFactory
 import scala.concurrent.duration._
